\name{hbrwted2a-package}
\alias{hbrwted2a-package}
\alias{hbrwted2a}
\docType{package}
\title{
\packageTitle{hbrwted2a}
}
\description{
\packageDescription{hbrwted2a}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hbrwted2a}
\packageIndices{hbrwted2a}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hbrwted2a}

Maintainer: \packageMaintainer{hbrwted2a}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
